+++ jklib/__init__.py
@@ -1,1 +1,2 @@
-"""Package that contains various utility functions spread across subpackages"""
+"""Package that contains various utility functions spread across
+subpackages."""