+++ jklib/db/__init__.py
@@ -1,1 +1,1 @@
-"""Utilities related to databases and ORMs"""
+"""Utilities related to databases and ORMs."""