+++ jklib/django_utils/db/__init__.py
@@ -1,1 +1,1 @@
-"""Utilities for the Django database and ORM"""
+"""Utilities for the Django database and ORM."""