+++ jklib/django_utils/db/managers.py
@@ -1,17 +1,23 @@
-"""Custom managers for Django models"""
+"""Custom managers for Django models."""
 
 
 # Built-in
-from typing import List
+from typing import Iterable, List, Optional, Sequence
 
 # Django
 from django.db import models
-from django.db.models import Model
 
 
-class NoBulkCreateManager(models.Manager):
-    """Prevents the use of the bulk_create method"""
+class NoBulkManager(models.Manager):
+    def bulk_create(
+        self,
+        objs: Iterable,
+        batch_size: Optional[int] = ...,
+        ignore_conflicts: bool = ...,
+    ) -> List:
+        raise NotImplementedError
 
-    def bulk_create(self, objs: List[Model], **kwargs) -> None:
-        """Overrides the 'bulk_create' method to make it non-usable"""
-        raise NotImplementedError("Cannot use bulk_create on this model")
+    def bulk_update(
+        self, objs: Iterable, fields: Sequence[str], batch_size: Optional[int] = ...
+    ) -> int:
+        raise NotImplementedError