+++ jklib/django_utils/db/queries.py
@@ -1,45 +1,43 @@
-"""Functions for performing SQL queries"""
+"""Functions for performing SQL queries."""
 
 # Built-in
-from typing import Optional, Type
+from typing import Any, List, Optional, Type
 
 # Django
-from django.db.models import Model, Q, QuerySet
+from django.db.models import Manager, Model
 
 
-def maybe_get_instance(model: Type[Model], *args, **kwargs) -> Optional[Model]:
-    """Tries to get an object in the database or returns None"""
+def maybe_get_instance(
+    model_class: Type[Model], *args: Any, **kwargs: Any
+) -> Optional[Model]:
+    """Tries to get an object in the database or returns None."""
     try:
-        item = model.objects.get(*args, **kwargs)
-    except model.DoesNotExist:
+        item = model_class.objects.get(*args, **kwargs)
+    except model_class.DoesNotExist:
         return None
     return item
 
 
-def filter_on_text(
-    queryset: QuerySet, searched_text: str, min_length: int, *fields
-) -> QuerySet:
-    """Filters a queryset by searching for a text value in different fields, using OR logic"""
-    if len(searched_text) >= min_length:
-        q = Q()
-        for field in fields:
-            key = f"{field}__icontains"
-            param = {key: searched_text}
-            new_q = Q(**param)
-            q |= new_q
-        queryset = queryset.filter(q)
-    return queryset
+def update_model(instance: Model, **kwargs: Any) -> Model:
+    """Updates multiple fields of a model instance and saves it."""
+    for key, value in kwargs.items():
+        setattr(instance, key, value)
+    instance.save()
+    return instance
 
 
-def single_sort_by(
-    queryset: QuerySet, sort_by: str = None, ascending: bool = False
-) -> QuerySet:
-    """
-    Orders and returns the queryset using the GET parameters of a request
-    The function takes care of checking if the "sort_by" key exists in the dict
-    """
-    if sort_by:
-        if not ascending:
-            sort_by = "-" + sort_by
-        queryset = queryset.order_by(sort_by)
-    return queryset
+def update_m2m(
+    m2m_field: Manager,
+    ids: List[str],
+) -> None:
+    """Updates a many-to-many field with a list of ids."""
+    unique_ids = set(ids or [])
+    existing_m2m_fks = m2m_field.all().values_list("id", flat=True)
+    # Delete m2m instances that are not in the provided ids
+    m2m_fks_to_delete = [id_ for id_ in existing_m2m_fks if id_ not in unique_ids]
+    if len(m2m_fks_to_delete) > 0:
+        m2m_field.remove(*m2m_fks_to_delete)
+    # Create m2m instances that are not in the existing instances
+    fks_to_add = [id_ for id_ in unique_ids if id_ not in existing_m2m_fks]
+    if len(fks_to_add) > 0:
+        m2m_field.add(*fks_to_add)