+++ jklib/django_utils/utils/templates.py
@@ -1,4 +1,4 @@
-"""Functions for managing templates in django"""
+"""Functions for managing templates in django."""
 
 # Built-in
 from typing import Dict
@@ -8,7 +8,7 @@ from django.template import loader
 
 
 def render_template(template_path: str, context: Dict) -> str:
-    """Renders an HTML page using a given template and a given context"""
+    """Renders an HTML page using a given template and a given context."""
     template = loader.get_template(template_path)
     rendered = template.render(context)
     return rendered