+++ jklib/dtsci/__init__.py
@@ -1,1 +1,2 @@
-"""Package with utility functions related to data-science (pandas, numpy, etc.)"""
+"""Package with utility functions related to data-science (pandas, numpy,
+etc.)"""